Use a bigger keyspace for the primary key in mysql/mssql
Also NOT NULL the mysql primary key (why was it nullable?)

The default signed integer primary key is pretty easy to max
out if you do a lot of locking over a long period of time.

Use a bigger keyspace and we get more time!

This doesn't solve the underlying issue which is that the primary
key should be reset every once in awhile probably.